requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
PyNaCl==1.5.0
yt-dlp>=2024.1.1
requests==2.31.0
#epic-free-games==0.1.8
